+++ cpp/runtimes/cluster-controller-runtime/JoynrClusterControllerRuntime.cpp
@@ -865,6 +865,7 @@ JoynrClusterControllerRuntime::~JoynrClusterControllerRuntime()
 
     multicastMessagingSkeletonDirectory->unregisterSkeleton<system::RoutingTypes::MqttAddress>();
 
+    messageSender->registerDispatcher(nullptr);
     if (joynrDispatcher != nullptr) {
         JOYNR_LOG_TRACE(logger, "joynrDispatcher");
         delete joynrDispatcher;